GetSemanticVersion is common to all OS
